allow players to url hack their own game IDs
This is somewhat tricky to reason about and hacky to implement. If the user provides a game ID in the URL by hand without clicking the new game button, how do we know if they're trying to join a game, or if they're trying to start a new one?

With this solution, we wait 4 seconds to try to connecto to other players in the same game; if none are found after 4 seconds, we assume they are trying to start a new game.